+++ djangae/contrib/backup/tests/test_tasks.py
@@ -3,7 +3,7 @@ import unittest
 from djangae.contrib import sleuth
 from djangae.contrib.backup.tasks import (
     AUTH_SCOPES,
-    _get_valid_export_models,
+    _get_valid_export_kinds,
     backup_datastore,
 )
 from djangae.test import TestCase
@@ -35,7 +35,7 @@ class GetValidExportModelsTestCase(TestCase):
     @override_settings(DJANGAE_BACKUP_EXCLUDE_MODELS=['backup_logentry'])
     @sleuth.switch('django.apps.apps.get_models', mock_get_app_models)
     def test_models_filtered(self):
-        valid_models = _get_valid_export_models(
+        valid_models = _get_valid_export_kinds(
             ['backup_logentry', 'backup_mockuser']
         )
         self.assertNotIn('backup_logentry', valid_models)
@@ -44,11 +44,38 @@ class GetValidExportModelsTestCase(TestCase):
     @override_settings(DJANGAE_BACKUP_EXCLUDE_APPS=['django'])
     @sleuth.switch('django.apps.apps.get_models', mock_get_app_models)
     def test_apps_filtered(self):
-        valid_models = _get_valid_export_models(
+        valid_models = _get_valid_export_kinds(
             ['backup_logentry', 'backup_mockuser']
         )
         self.assertIn('backup_mockuser', valid_models)
         self.assertNotIn('django_admin_log', valid_models)
+
+    def test_models_using_the_same_table_only_listed_once(self):
+        class ModelFoo(models.Model):
+            class Meta:
+                db_table = "foo"
+
+        class ModelBar(models.Model):
+            class Meta:
+                db_table = "foo"
+
+        def mock_get_app_models(**kwargs):
+            return [ModelFoo, ModelBar]
+
+        with sleuth.switch('django.apps.apps.get_models', mock_get_app_models):
+            valid_models = _get_valid_export_kinds()
+            self.assertEquals(['foo'], valid_models)
+
+    @override_settings(DJANGAE_BACKUP_EXCLUDE_MODELS=['backup_logentry'])
+    @sleuth.switch('django.apps.apps.get_models', mock_get_app_models)
+    def test_kinds_are_deduplicated(self):
+        valid_models = _get_valid_export_kinds(kinds=[
+            'backup_logentry',
+            'backup_logentry',
+            'backup_mockuser',
+            'backup_mockuser',
+        ])
+        self.assertEquals(['backup_mockuser'], valid_models)
 
 
 class BackupTestCase(TestCase):